Updated config files to more reasonable filenames
