+++ protractor.conf.browserstack-remote.js
@@ -1,9 +1,8 @@
 // Protractor configuration file, see link for more information
 // https://github.com/angular/protractor/blob/master/lib/config.ts
-// Note that this is the config file for e2e tests using BrowserStack's automation API (against local hosts). See protractor.conf.browserstack-remote.js for remote-test settings.
+// Note that this is the config file for e2e tests using BrowserStack's automation API (against remote hosts). See protractor.conf.browserstack-local.js for local-test settings.
 
 var BROWSERSTACK_CREDENTIALS = require('./e2e/browserstack-credentials');
-var browserstack = require('browserstack-local');
 
 exports.config = {
   allScriptsTimeout: 11000,
@@ -29,8 +28,7 @@ exports.config = {
     'browserstack.key' : BROWSERSTACK_CREDENTIALS.KEY,
     'browserstack.debug' : true,
     'browserstack.video' : true,
-    'browserstack.local' : true,
-    'acceptSslCerts' : true,
+    'acceptSslCerts' : true
   },
 
   'multiCapabilities': [{
@@ -48,29 +46,7 @@ exports.config = {
     'os' : 'OS X',
     'os_version' : 'Mojave',
     'resolution' : '1600x1200'
-  }],
-
-    // Code to start browserstack local before start of test
-  beforeLaunch: function(){
-    console.log("Connecting local");
-    return new Promise(function(resolve, reject){
-      exports.bs_local = new browserstack.Local();
-      exports.bs_local.start({'key': exports.config.commonCapabilities['browserstack.key'] }, function(error) {
-        if (error) return reject(error);
-        console.log('Connected. Now testing...');
-
-        resolve();
-      });
-    });
-  },
-
-  // Code to stop browserstack local after end of test
-  afterLaunch: function(){
-    return new Promise(function(resolve, reject){
-      exports.bs_local.stop(resolve);
-    });
-  }
-
+  }]
 };
 
 // Code to support common capabilities
