+++ test.sh
@@ -43,14 +43,14 @@ run_protractor() {
         # Source BrowserStack credentials
         if [ "$host" = 1 ]; then
             echo "Running BrowserStack tests against remote host: $1"
-            ./node_modules/.bin/protractor protractor.conf.js --baseUrl  $1
+            ./node_modules/.bin/protractor protractor.conf.browserstack-remote --baseUrl  $1
         else
             echo "Running BrowserStack tests against local host: $1"
             ./node_modules/.bin/protractor protractor.conf.browserstack-local --baseUrl  $1
         fi
     else
         echo "Running Local Selenium tests against host: $1"
-        ./node_modules/.bin/protractor protractor_selenium.conf.js --baseUrl  $1
+        ./node_modules/.bin/protractor protractor.conf.js --baseUrl  $1
     fi
 }
